feat: APP-670 change folder ecocreditsbyaccount to profile, routes update
