--- web-marketplace/src/pages/EcocreditsByAccount/index.ts
@@ -1,1 +1,0 @@
-export { EcocreditsByAccount as default } from './EcocreditsByAccount';
