+++ web-marketplace/src/pages/Profile/index.ts
@@ -1,0 +1,1 @@
+export { Profile as default } from './Profile';
