+++ web-marketplace/src/pages/index.ts
@@ -10,7 +10,6 @@ export * from './CreateCreditClass';
 export * from './CreditClassDetails';
 export * from './Dashboard';
 export * from './Description';
-export * from './EcocreditsByAccount';
 export * from './Eligibility';
 export * from './Home';
 export * from './LandStewards';
@@ -19,6 +18,7 @@ export * from './Media';
 export * from './MethodologiesList';
 export * from './MethodologyDetails';
 export * from './NotFound';
+export * from './Profile';
 export * from './Project';
 export * from './ProjectEdit';
 export * from './ProjectLocation';
